requirements.txt
pillow>=10.0.0
exifread>=3.0.0
opencv-python>=4.8.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
geopy>=2.3.0
openai>=1.0.0
requests>=2.31.0
click>=8.1.0
tqdm>=4.65.0
python-dateutil>=2.8.0
transformers>=4.20.0
numpy>=1.21.0,<2.0.0
pytest>=7.0.0
python-dotenv>=1.0.0

# Optional packages (require additional setup)
# Install first: brew install cmake && brew install dlib
face-recognition>=1.3.0